TINY-9253: test refactor
